+++ src/test/java/org/folio/rest/api/StorageTestSuite.java
@@ -91,7 +91,8 @@ import org.junit.runners.Suite;
   RetainLeadingZeroesMigrationScriptTest.class,
   StatisticalCodeTest.class,
   UpcIsmnMigrationScriptTest.class,
-  InstanceStorageInstancesBulkApiTest.class
+  InstanceStorageInstancesBulkApiTest.class,
+  RemoveInstanceShadowCopiesWithoutSharedRecordTest.class
 
   // These fail.
   //ReferenceTablesTest.class,